fix(js): console utf8 output

+++ src/shell/script/binding_types.h
@@ -258,7 +258,7 @@ struct js_menu_data {
   std::optional<bool> disabled;
   // 仅作为信息标识，不影响行为
   // Only for information, set this changes nothing
-  std::optional<size_t> wID;
+  std::optional<int64_t> wID;
 };
 
 struct menu_item_controller {
